Changed to Long instead of Integer. Dont forget to change student id, courses id, and all 3 rows in the student_courses table to BIGINT instead of INT.
Also been trying to get delete to work but cant seem to find why it doesnt.

+++ GritAcademyPortal/src/main/java/com/GritAcademyPortal/courses/CoursesController.java
@@ -35,7 +35,7 @@ public class CoursesController {
     }
 
     @GetMapping(value = "/coursesbyidwithstudents/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<CoursesDTO>> getCoursesByIdWithStudents(@PathVariable(value = "id") Integer id) {
+    public ResponseEntity<List<CoursesDTO>> getCoursesByIdWithStudents(@PathVariable(value = "id") Long id) {
         List<CoursesDTO> coursesDTO = coursesService.getCoursesByIdWithStudents(id);
         return new ResponseEntity<>(coursesDTO, HttpStatus.OK);
     }
