+++ GritAcademyPortal/src/main/java/com/GritAcademyPortal/studentsCourses/StudentsCoursesRepository.java
@@ -2,5 +2,5 @@ package com.GritAcademyPortal.studentsCourses;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentsCoursesRepository extends JpaRepository<StudentsCourses, Integer> {
+public interface StudentsCoursesRepository extends JpaRepository<StudentsCourses, Long> {
 }
